refactor: move OpenAPI types into openapi/ folder, rename emitter
- OpenAPI document types and parser live in src/openapi/
- IR + codegen pipeline types stay in src/types.ts
- Rename openapi.ts → emitter.ts (it consumes the IR, not OpenAPI)
- Remove ProtocolGenerator indirection

Made-with: Cursor
