+++ packages/codegen/src/index.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2026 Reactor Technologies, Inc. All rights reserved.
 
 export { generateModelSdk, writePackage, loadSchema, parseSchema } from "./codegen.js";
+export type { OpenApiSchema } from "./openapi/index.js";
 export type {
-  OpenApiSchema,
   ModelSchema,
   CodegenOptions,
   EventSchema,
