+++ packages/codegen/src/openapi/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) 2026 Reactor Technologies, Inc. All rights reserved.
+
+export { loadSchema, parseSchema } from "./parser.js";
+export type { OpenApiSchema } from "./types.js";
